+++ packages/admin-rte/src/core/Controls/FeaturesButtonGroup.tsx
@@ -8,25 +8,26 @@ import * as sc from "./FeaturesButtonGroup.sc";
 
 interface IProps {
     features: IFeatureConfig[];
+    disabled?: boolean;
 }
 
-export default function FeaturesButtonGroup({ features }: IProps) {
+export default function FeaturesButtonGroup({ features, disabled: globallyDisabled }: IProps) {
     if (!features.length) {
         return null;
     }
 
     return (
         <Box display="inline-flex" justifyContent="flex-start">
-            {features.map(({ name, onButtonClick, tooltipText, ...rest }) => (
+            {features.map(({ name, onButtonClick, tooltipText, disabled, ...rest }) => (
                 <sc.ButtonWrapper key={name}>
                     {tooltipText ? (
                         <Tooltip title={tooltipText} placement="top">
                             <span>
-                                <ControlButton onButtonClick={onButtonClick} {...rest} />
+                                <ControlButton onButtonClick={onButtonClick} disabled={globallyDisabled || disabled} {...rest} />
                             </span>
                         </Tooltip>
                     ) : (
-                        <ControlButton onButtonClick={onButtonClick} {...rest} />
+                        <ControlButton onButtonClick={onButtonClick} disabled={globallyDisabled || disabled} {...rest} />
                     )}
                 </sc.ButtonWrapper>
             ))}