+++ packages/admin-rte/src/core/Controls/HistoryContols.tsx
@@ -4,10 +4,10 @@ import { IControlProps } from "../types";
 import FeaturesButtonGroup from "./FeaturesButtonGroup";
 import useHistory from "./useHistory";
 
-export default function HistoryControls({ editorState, setEditorState, options }: IControlProps) {
+export default function HistoryControls({ editorState, setEditorState, options, disabled }: IControlProps) {
     const { features } = useHistory({ editorState, setEditorState, supportedThings: options.supports });
     if (features.length < 1) {
         return null;
     }
-    return <FeaturesButtonGroup features={features} />;
+    return <FeaturesButtonGroup features={features} disabled={disabled} />;
 }