+++ packages/admin-rte/src/core/Controls/ListsIndentControls.tsx
@@ -4,10 +4,15 @@ import { IControlProps } from "../types";
 import FeaturesButtonGroup from "./FeaturesButtonGroup";
 import useListIndent from "./useListIndent";
 
-export default function ListsIndentControls({ editorState, setEditorState, options: { supports: supportedThings, listLevelMax } }: IControlProps) {
+export default function ListsIndentControls({
+    editorState,
+    setEditorState,
+    options: { supports: supportedThings, listLevelMax },
+    disabled,
+}: IControlProps) {
     const { features } = useListIndent({ editorState, setEditorState, supportedThings, listLevelMax });
     if (features.length < 1) {
         return null;
     }
-    return <FeaturesButtonGroup features={features} />;
+    return <FeaturesButtonGroup features={features} disabled={disabled} />;
 }