+++ packages/admin-stories/src/admin-rte/Field.tsx
@@ -1,4 +1,4 @@
-import { Field } from "@comet/admin";
+import { Field, FinalFormInput } from "@comet/admin";
 import { createFinalFormRte } from "@comet/admin-rte";
 import { Button, Typography } from "@material-ui/core";
 import { storiesOf } from "@storybook/react";
@@ -9,6 +9,7 @@ const { RteField, RteReadOnly } = createFinalFormRte();
 
 function Story() {
     const [submitedValue, setSubmittedValue] = React.useState<{ rteContent: any }>({ rteContent: undefined });
+    const [disabled, toggleDisabled] = React.useReducer((s) => !s, false);
 
     return (
         <div>
@@ -16,11 +17,22 @@ function Story() {
                 <Form
                     onSubmit={(values: { rteContent: any }) => {
                         //
-                        setSubmittedValue({ rteContent: values });
+                        setSubmittedValue({ rteContent: values.rteContent });
                     }}
                     render={({ handleSubmit }) => (
                         <form onSubmit={handleSubmit}>
-                            <Field name="rteContent" label="Rich Text" component={RteField} />
+                            <Field name="rteContent" label="Rich Text" component={RteField} disabled={disabled} />
+                            <Field name="somenthingElse" label="Something else" component={FinalFormInput} disabled={disabled} />
+                            <Button
+                                color="secondary"
+                                variant="contained"
+                                type="button"
+                                component={"button"}
+                                disableTouchRipple
+                                onClick={toggleDisabled}
+                            >
+                                <Typography variant="button">{disabled ? "Enable" : "Disable"} inputs</Typography>
+                            </Button>
                             <Button color="primary" variant="contained" type="submit" component={"button"} disableTouchRipple>
                                 <Typography variant="button">Submit</Typography>
                             </Button>