+++ packages/admin-rte/CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file. This projec
 
 -   Add default styles (MUI) to built-in blocktypes
 -   Make built-in blocktypes styleable
+-   Supports disabled-attribute
 
 ### Internal Changes